Add support for resizing render targets and swapchain.
This is a bit of a "hack", but it works. So we'll leave it as is for now and take another look later.
The main issue here is that our systems are async and executed in parallel at all times, even parts of the renderer. So the only reliable way to communicate is events that will be read the next frame.